+++ src/import.js
@@ -1,31 +1,5 @@
+import dogui_vue from './vue/interface';
 export default {
-    title: require('./vue/out/title.vue'),
-    block: require('./vue/out/block.vue'),
-    cell: require('./vue/out/cell.vue'),
-    button: require('./vue/out/button.vue'),
-    search: require('./vue/out/searchBar.vue'),
-    upload: require('./vue/out/picture.vue'),
-    input: require('./vue/out/input.vue'),
-    select: require('./vue/out/selector.vue'),
-    info: require('./vue/out/info.vue'),
-    view: require('./vue/out/pictureView.vue'),
-    textarea: require('./vue/out/textarea.vue'),
-    cover: require('./vue/out/cover.vue'),
-    navBar: require('./vue/out/navbar.vue')
+    vue: dogui_vue
 }
-export const pc = {
-
-}
-export const title = require('./vue/out/title.vue');
-export const block = require('./vue/out/block.vue');
-export const cell = require('./vue/out/cell.vue');
-export const button = require('./vue/out/button.vue');
-export const search = require('./vue/out/searchBar.vue');
-export const upload = require('./vue/out/picture.vue');
-export const input = require('./vue/out/input.vue');
-export const select = require('./vue/out/selector.vue');
-export const info = require('./vue/out/info.vue');
-export const view = require('./vue/out/pictureView.vue');
-export const textarea = require('./vue/out/textarea.vue');
-export const cover = require('./vue/out/cover.vue');
-export const navBar = require('./vue/out/navbar.vue');
+export const vue = dogui_vue;