+++ src/main.js
@@ -3,6 +3,11 @@
 // This is done with the browser options. For the config, see package.json
 import Vue from 'vue'
 import App from './App.vue'
+import {
+  vue
+} from './import'
+
+Vue.use(vue);
 
 new Vue({ // eslint-disable-line no-new
   el: '#app',