+++ src/vue/import.js
@@ -1,0 +1,27 @@
+import dog_title from './out/title.vue';
+import dog_block from './out/block.vue';
+import dog_cell from './out/cell.vue';
+import dog_button from './out/button.vue';
+import dog_search from './out/searchBar.vue';
+import dog_upload from './out/picture.vue';
+import dog_input from './out/input.vue';
+import dog_select from './out/selector.vue';
+import dog_info from './out/info.vue';
+import dog_view from './out/pictureView.vue';
+import dog_textarea from './out/textarea.vue';
+import dog_cover from './out/cover.vue';
+import dog_navBar from './out/navbar.vue';
+
+export const title = dog_title;
+export const block = dog_block;
+export const cell = dog_cell;
+export const button = dog_button;
+export const search = dog_search;
+export const upload = dog_upload;
+export const input = dog_input;
+export const select = dog_select;
+export const info = dog_info;
+export const view = dog_view;
+export const textarea = dog_textarea;
+export const cover = dog_cover;
+export const navBar = dog_navBar;